+++ Changelog.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v1.5.11] - next
 
+Major AFCL update
+
 ### Added
 - AFCL Table component
+- AFCL Chart components and documentation
 
 ## [v1.5.10]